utils/docker-run: fix running when CWD is not MAIN_DIR
Commit 08edccebcb92 (utils/docker-run: fix support for git-worktrees)
got last-minute changes when it was applied, and the case when the
current working directory is not the top of the current working copy
got broken.

Fix that by duplicating (and thus reinstating) the 'cd MAIN_DIR' to
match what is done when retrieving the git-common-dir.

Fixes: 08edccebcb92

Reported-by: Brandon Maier <Brandon.Maier@collins.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit 0407d34962f120d40dab61c042a737e9e0daf574)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>