package/host-gdb: disable source-highlight
This is the follow-up commit to the ca6c593 which disables
sourcehighlight dependency for the target GDB package.
Do the same for the host variant to be sure that this
won't be picked up from the host system.

Signed-off-by: Bartosz Bilas <b.bilas@grinn-global.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit 0ae981ac1138e93aeb5a647bf16fdccff96a6ea1)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>