package/libmodplug: use a full-length hash as version
Commit c9778e00e92c (package/libmodplug: update to git version)
improperly used a short hash as version, so switch to the full-length
hash.

Github use the full-length hash when it creates the top-level directory
of the generated archive, so the hash of the archive does not in fact
change, only its filename. This is perfectly fine fine our handling of
s.b.o.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit 0b68c23b9b29380d05d9608985a9904795e93ec9)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>