configs/stm32mp157c_odyssey_defconfig: use a fixed TF-A version
Commit 81bcc03b3b3 (configs/stm32mp157c_odyssey: new defconfig) forgot to
specify a fixed TF-A version, so do that now.

When the defconfig was added, the default version was v2.5 - So use that.
Similarly to the other stm32mp1 defconfigs, this needs disabling -Werror
with E=0 to fix a build issue with GCC >= 12.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit 42779abbda975530f40ddd00c29035ff24e03ee1)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>